+++ src/caselawclient/xquery_type_dicts.py
@@ -101,6 +101,7 @@ class GetVersionCreatedDict(MarkLogicAPIDict):
 
 # insert_document.xqy
 class InsertDocumentDict(MarkLogicAPIDict):
+    annotation: str
     document: str
     uri: MarkLogicDocumentURIString